+++ community/.gitignore
@@ -18,9 +18,6 @@ target
 cypher/project
 cypher/lib_managed
 cypher/.cache
-server/src/main/resources/webadmin-html/js/neo4j/
-server/src/main/resources/webadmin-html/js/ribcage/
-server/src/main/resources/webadmin-html/js/webadmin.js
 artifacts
 */bin
 */pom-with-build-number.xml
