+++ community/server/src/main/java/org/neo4j/server/AbstractNeoServer.java
@@ -19,10 +19,6 @@
  */
 package org.neo4j.server;
 
-import com.sun.jersey.api.core.HttpContext;
-import org.apache.commons.configuration.Configuration;
-import org.bouncycastle.operator.OperatorCreationException;
-
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
@@ -35,6 +31,10 @@ import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.regex.Pattern;
 import javax.servlet.Filter;
+
+import com.sun.jersey.api.core.HttpContext;
+import org.apache.commons.configuration.Configuration;
+import org.bouncycastle.operator.OperatorCreationException;
 
 import org.neo4j.bolt.security.ssl.Certificates;
 import org.neo4j.bolt.security.ssl.KeyStoreFactory;
@@ -85,6 +85,7 @@ import org.neo4j.udc.UsageData;
 import static java.lang.Math.round;
 import static java.lang.String.format;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
+
 import static org.neo4j.helpers.Clock.SYSTEM_CLOCK;
 import static org.neo4j.helpers.collection.Iterables.map;
 import static org.neo4j.kernel.impl.util.JobScheduler.Groups.serverTransactionTimeout;
@@ -105,7 +106,6 @@ public abstract class AbstractNeoServer implements NeoServer
 
     private static final Pattern[] DEFAULT_URI_WHITELIST = new Pattern[]{
             Pattern.compile( "/browser.*" ),
-            Pattern.compile( "/webadmin.*" ),
             Pattern.compile( "/" )
     };
