--- community/server/src/test/java/org/neo4j/server/webadmin/rest/VersionAndEditionServiceTest.java
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2002-2016 "Neo Technology,"
- * Network Engine for Objects in Lund AB [http://neotechnology.com]
- *
- * This file is part of Neo4j.
- *
- * Neo4j is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.neo4j.server.webadmin.rest;
-
-import org.junit.Test;
-
-import org.neo4j.server.CommunityNeoServer;
-import org.neo4j.server.rest.management.VersionAndEditionService;
-
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.mock;
-
-public class VersionAndEditionServiceTest
-{
-    @Test
-    public void shouldReturnReadableStringForServiceName() throws Exception
-    {
-        // given
-        VersionAndEditionService service = new VersionAndEditionService( mock( CommunityNeoServer.class ) );
-
-        // when
-        String serviceName = service.getName();
-        // then
-        assertEquals( "version", serviceName );
-    }
-
-    @Test
-    public void shouldReturnSensiblePathWhereServiceIsHosted() throws Exception
-    {
-        // given
-        VersionAndEditionService service = new VersionAndEditionService( mock( CommunityNeoServer.class ) );
-
-        // when
-        String serverPath = service.getServerPath();
-
-        // then
-        assertEquals( "server/version", serverPath );
-    }
-}
-
-
