+++ enterprise/server-enterprise/src/test/java/org/neo4j/server/rest/MasterInfoServiceTest.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.server.webadmin.rest;
+package org.neo4j.server.rest;
 
 import javax.ws.rs.core.Response;
