+++ manual/contents/src/tools/index.asciidoc
@@ -13,9 +13,5 @@ include::{importdir}/neo4j-import-tool-docs-jar/ops/import-tool.asciidoc[]
 
 :leveloffset: 1
 
-include::webadmin.asciidoc[]
-
-:leveloffset: 1
-
 include::{importdir}/neo4j-shell-docs-jar/dev/shell.asciidoc[]
