--- manual/contents/src/tools/webadmin.asciidoc
@@ -1,10 +1,0 @@
-[[tools-webadmin]]
-= Web Interface =
-
-The Neo4j Web Interface is the primary user interface for Neo4j.
-
-The tool is available at http://127.0.0.1:7474/ after you have installed the <<server-installation,Neo4j Server>>.
-
-See the tool itself for more information!
-
-
