+++ manual/cypher/cypher-docs/src/docs/dev/tutorials/cypher.asciidoc
@@ -7,8 +7,8 @@ In the online edition of this manual, all queries in this section can be execute
 
 Otherwise, first get the Neo4j server running to try things out locally.
 Instructions are found in <<server-installation>>.
-With the server running, you can choose to issue Cypher queries from either the web interface or the Neo4j shell.
-See <<tools-webadmin>> or <<shell>>.
+With the server running, you can choose to issue Cypher queries from either the Neo4j Browser or the Neo4j shell.
+See <<shell>>.
 
 :leveloffset: 2
