+++ packaging/standalone/document.properties
@@ -4,5 +4,3 @@ shell.fullname=Neo4j Shell
 shell.shortname=neo4j-shell
 libs.fullname=Neo4j Libraries
 libs.shortname=neo4j-libs
-webadmin.fullname=Neo4j Web Administration
-webadmin.shortname=Webadmin
